feat: add update route usecase
